[cleanup] Remove unnecessary mcast tests (#92)
The exclusion/prune and no_admin_ula mcast integration tests were
unnecessary  and part of previously configurable code that was made
obsolete during the last review cycle of the large mcast PR.

The exclusion/prune test was being flaky on CI, but didn't set up proper
MAC linking anyway. It's not useful because we force 0 for the exclusion
flag when we configure replication on the underlay groups.

Includes:
  * use oxnet's derive mac facility for tests, updating the dep
  * mcast integration test comment cleanup
  * move P4 # comment to //