[api] New response types for External/Underlay groups, rename fns, consistentcy
This includes `MulticastUnderlayGroupResponse` and `MulticastExternalGroupResponse`, and a
unified response type for lists, mixed result calls `MulticastGroupResponse`. We also added
an AdminScoped type for underlay and consistent naming throughout. We
also rename structs for consistency, and handle rollback at the
boundary calls to internal fns.

This PR has been updated to accomodate the new API trait,
https://github.com/oxidecomputer/omicron/issues/8922, so it adjusts
a lot from the previous code and commit.