remove listen_addr requirement for rpw startup (#117)
We were requiring non-localhost listen_addrs for NAT RPW startup, and there isn't really a good reason to do so at this point. Requiring non-localhost listen addrs means that the RPW will never run in the omicron test suite, since the test suite exclusively uses localhost listen addrs.

We also allow the optional injection of mgs_addr so that we can get dendrite working properly in the test suite. Normally we can just listen on localhost:[MGS_PORT] but once again, the omicron test suite relies on randomized ports so it can run multiple tests in parallel, so we can't exclusively rely on that pattern if we want dendrite to work properly in the omicron test environment.