sync dhcpv6 addresses to ASIC
This PR adds a DHCPv6 address synchronizer to tfportd. This
synchronizier checks both techport interfaces for DHCPv6 addresses every
30 seconds (same interval as the SLAAC synchronizer). If a DHCPv6
address is found on a techport, it is synchronized to the ASIC local
v6 address table.

Initial testing shows that this is all we need to support DHCPv6 for
techport interfaces. This was tested on a racklette in the lab whose
technician ports were connected to a management switch set up with a
DHCPv6 server configuration.

Existing P4 plumbing for link-local multicast is sufficient to transit
the required ICMPv6 router advertisement messaging and DHCPv6 messaging
between FF02:1:2 and link-local addresses to and from the switch zone.

The end result is a racklette with a DHCPv6 address that is reachable
from the management switch.